Added sim65 code version that fixes all known issues.

+++ 6502-fixed.h
@@ -32,11 +32,10 @@
 /*****************************************************************************/
 
 
-
 #ifndef _6502_H
 #define _6502_H
 
-
+#include <inttypes.h>
 
 /*****************************************************************************/
 /*                                   Data                                    */
@@ -57,13 +56,13 @@ extern CPUType CPU;
 /* 6502 CPU registers */
 typedef struct CPURegs CPURegs;
 struct CPURegs {
-    unsigned    AC;             /* Accumulator */
-    unsigned    XR;             /* X register */
-    unsigned    YR;             /* Y register */
-    unsigned    ZR;             /* Z register */
-    unsigned    SR;             /* Status register */
-    unsigned    SP;             /* Stackpointer */
-    unsigned    PC;             /* Program counter */
+    uint8_t     AC;             /* Accumulator */
+    uint8_t     XR;             /* X register */
+    uint8_t     YR;             /* Y register */
+    uint8_t     ZR;             /* Z register */
+    uint8_t     SR;             /* Status register */
+    uint8_t     SP;             /* Stackpointer */
+    uint16_t    PC;             /* Program counter */
 };
 
 /* Status register bits */
